completed FEC metadata handling (no custom yet).. need to bug test KSK's and SSK'2
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/contrib@4495 67a373e5-eb02-0410-a15c-ee090a768436

+++ fcptools/ezFCPlib/ezFCPlib.h
@@ -211,7 +211,7 @@ typedef struct {
 	char  fec_algorithm[L_KEY+1];
  
 	int   filelength;
-	int   offset;
+	long  offset;
 	int   block_count;
 	int   block_size;
 	int   datablock_offset;
@@ -333,6 +333,7 @@ typedef struct {
 
 	int    filelength;
 	long   offset;
+
 	int    block_count;
 	int    block_size;
 	int    datablock_offset;
